chore: move ioc container out of the application layer
