+++ inversify-generator.json
@@ -1,4 +1,4 @@
 {
-  "output": "./src/shared/application/ioc/__generated__",
+  "output": "./src/shared/ioc/__generated__",
   "binding": "dynamic"
 }
