+++ src/posts/presentation/pages/posts/posts.tsx
@@ -1,9 +1,9 @@
 import css from "@/src/shared/presentation/styles/wrapper.css";
 import { SimpleCard } from "@/src/shared/presentation/components/simple-card/simple-card";
-import { locator } from "@/src/shared/application/ioc/__generated__";
-import type { IocProvider } from "@/src/shared/application/ioc/interfaces";
+import { locator } from "@/src/shared/ioc/__generated__";
+import type { IocProvider } from "@/src/shared/ioc/interfaces";
 import type { GetPostsUseCase } from "@/src/posts/domain/use-cases/get-posts-use-case";
-import { TYPES } from "@/src/shared/application/ioc/__generated__/types";
+import { TYPES } from "@/src/shared/ioc/__generated__/types";
 import type { Post } from "@/src/posts/domain/models/post";
 import { notFound } from "next/navigation";
 import { getTranslations } from "next-intl/server";
