+++ src/shared/ioc/interfaces.ts
