+++ src/users/data/repositories/users-repository.ts
@@ -1,8 +1,8 @@
-import { TYPES } from "@/src/shared/application/ioc/__generated__/types";
+import { TYPES } from "@/src/shared/ioc/__generated__/types";
 import { inject, injectable } from "inversify";
 import type { User } from "@/src/users/domain/models/user";
 import { plainToClass } from "class-transformer";
-import type { IocProvider } from "@/src/shared/application/ioc/interfaces";
+import type { IocProvider } from "@/src/shared/ioc/interfaces";
 import { UserDTO } from "@/src/users/data/dtos/user";
 import type { IUsersRepository } from "@/src/users/domain/interfaces/users-repository";
 import type { RestService } from "@/src/shared/data/services/rest-service";
