+++ src/users/domain/use-cases/get-users-use-case.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from "inversify";
 import type { User } from "@/src/users/domain/models/user";
-import type { IocProvider } from "@/src/shared/application/ioc/interfaces";
-import { TYPES } from "@/src/shared/application/ioc/__generated__/types";
+import type { IocProvider } from "@/src/shared/ioc/interfaces";
+import { TYPES } from "@/src/shared/ioc/__generated__/types";
 import type { IUsersRepository } from "@/src/users/domain/interfaces/users-repository";
 
 @injectable()
