+++ src/users/presentation/pages/users-with-provider/users-with-provider.tsx
@@ -1,8 +1,8 @@
 import css from "@/src/shared/presentation/styles/wrapper.css";
-import { locator } from "@/src/shared/application/ioc/__generated__";
-import type { IocProvider } from "@/src/shared/application/ioc/interfaces";
+import { locator } from "@/src/shared/ioc/__generated__";
+import type { IocProvider } from "@/src/shared/ioc/interfaces";
 import type { GetUsersUseCase } from "@/src/users/domain/use-cases/get-users-use-case";
-import { TYPES } from "@/src/shared/application/ioc/__generated__/types";
+import { TYPES } from "@/src/shared/ioc/__generated__/types";
 import UsersList from "@/src/users/presentation/containers/users-list";
 import { instanceToPlain } from "class-transformer";
 import type { User } from "@/src/users/domain/models/user";
